feat(session): versioned schema migrations, busy_timeout, missing indexes
