+++ src/session/store.rs
@@ -1,14 +1,42 @@
+use std::collections::HashSet;
 use std::path::{Path, PathBuf};
+use std::sync::{Mutex, OnceLock};
+use std::time::Duration;
 
 use rusqlite::Connection;
 
 use super::context::StorageContext;
+use super::migrations;
 use crate::error::Result;
 
 /// Subdirectory of the workspace holding the session database.
 const DB_SUBDIR: &str = "session_db";
 /// Database filename inside [`DB_SUBDIR`].
 const DB_FILE: &str = "sessions.db";
+
+/// How long a statement waits for a competing writer's lock before giving up
+/// with `SQLITE_BUSY`.
+///
+/// SQLite's default is **zero**: a `BEGIN IMMEDIATE` that finds the write lock
+/// held fails instantly rather than waiting. Every claim/gate/sequence
+/// allocation in this module is written on the assumption that racing writers
+/// *serialize* at `BEGIN` — with no busy handler installed they do not, they
+/// just fail, and the caller sees a spurious storage error under ordinary
+/// concurrency. Five seconds is long enough to ride out any transaction this
+/// module takes (all of them are a handful of small statements) and short
+/// enough to surface a genuine deadlock rather than hang.
+const BUSY_TIMEOUT: Duration = Duration::from_secs(5);
+
+/// Databases whose migrations have already been applied **in this process**.
+///
+/// [`migrations::apply`] is idempotent and cheap when up to date (one indexed
+/// row read), but a connection is opened per operation, so even that read is
+/// worth skipping once we know the file is current. Keyed by resolved path;
+/// entries are only inserted after a successful migration run.
+fn migrated_paths() -> &'static Mutex<HashSet<PathBuf>> {
+    static MIGRATED: OnceLock<Mutex<HashSet<PathBuf>>> = OnceLock::new();
+    MIGRATED.get_or_init(|| Mutex::new(HashSet::new()))
+}
 
 /// Resolves the session database path for a workspace root.
 ///
@@ -38,9 +66,44 @@ pub fn with_connection<T>(
 
     let conn = Connection::open(&db_path)
         .storage_context(&format!("failed to open session DB: {}", db_path.display()))?;
+    prepare_connection(&conn)?;
 
-    init_schema(&conn)?;
+    // Migrations run once per database per process; see `migrated_paths`.
+    let already_migrated = {
+        let guard = migrated_paths()
+            .lock()
+            .map_err(|e| poisoned("migration cache", e))?;
+        guard.contains(&db_path)
+    };
+    if !already_migrated {
+        migrations::apply(&conn)?;
+        migrated_paths()
+            .lock()
+            .map_err(|e| poisoned("migration cache", e))?
+            .insert(db_path.clone());
+    }
+
     f(&conn)
+}
+
+fn poisoned(what: &str, err: impl std::fmt::Display) -> crate::error::TinyAgentsError {
+    crate::error::TinyAgentsError::Storage(format!("session DB {what} lock poisoned: {err}"))
+}
+
+/// Applies the per-connection pragmas every session-DB handle needs.
+///
+/// `journal_mode = WAL` is persistent (stored in the file header) but is set
+/// here so a freshly created database gets it; `foreign_keys` and
+/// `busy_timeout` are **per connection** and must be set on every open.
+fn prepare_connection(conn: &Connection) -> Result<()> {
+    conn.busy_timeout(BUSY_TIMEOUT)
+        .storage_context("failed to set session DB busy_timeout")?;
+    conn.execute_batch(
+        "PRAGMA journal_mode = WAL;
+         PRAGMA foreign_keys = ON;",
+    )
+    .storage_context("failed to apply session DB pragmas")?;
+    Ok(())
 }
 
 /// Opens the session database and runs `f` inside a single **immediate**
@@ -88,94 +151,7 @@ pub fn with_transaction<T>(
 pub fn with_memory_connection<T>(f: impl FnOnce(&Connection) -> Result<T>) -> Result<T> {
     let conn =
         Connection::open_in_memory().storage_context("failed to open in-memory session DB")?;
-    init_schema(&conn)?;
+    prepare_connection(&conn)?;
+    migrations::apply(&conn)?;
     f(&conn)
-}
-
-pub(super) fn init_schema(conn: &Connection) -> Result<()> {
-    conn.execute_batch(
-        "PRAGMA journal_mode = WAL;
-         PRAGMA foreign_keys = ON;
-
-         CREATE TABLE IF NOT EXISTS sessions (
-            id                    TEXT PRIMARY KEY,
-            agent_definition_id   TEXT NOT NULL,
-            agent_definition_name TEXT NOT NULL,
-            session_key           TEXT NOT NULL,
-            parent_session_id     TEXT,
-            thread_id             TEXT,
-            source_channel        TEXT,
-            status                TEXT NOT NULL DEFAULT 'running',
-            model                 TEXT,
-            turn_count            INTEGER NOT NULL DEFAULT 0,
-            input_tokens          INTEGER NOT NULL DEFAULT 0,
-            output_tokens         INTEGER NOT NULL DEFAULT 0,
-            cached_input_tokens   INTEGER NOT NULL DEFAULT 0,
-            cost_usd              REAL NOT NULL DEFAULT 0.0,
-            transcript_path       TEXT,
-            started_at            TEXT NOT NULL,
-            ended_at              TEXT,
-            FOREIGN KEY (parent_session_id) REFERENCES sessions(id) ON DELETE SET NULL
-         );
-         CREATE INDEX IF NOT EXISTS idx_sessions_agent ON sessions(agent_definition_id);
-         CREATE INDEX IF NOT EXISTS idx_sessions_status ON sessions(status);
-         CREATE INDEX IF NOT EXISTS idx_sessions_started ON sessions(started_at);
-         CREATE INDEX IF NOT EXISTS idx_sessions_parent ON sessions(parent_session_id);
-         CREATE INDEX IF NOT EXISTS idx_sessions_thread ON sessions(thread_id);
-         CREATE INDEX IF NOT EXISTS idx_sessions_channel ON sessions(source_channel);
-         CREATE INDEX IF NOT EXISTS idx_sessions_key ON sessions(session_key);
-
-         CREATE TABLE IF NOT EXISTS session_messages (
-            id          INTEGER PRIMARY KEY AUTOINCREMENT,
-            session_id  TEXT NOT NULL,
-            role        TEXT NOT NULL,
-            content     TEXT NOT NULL,
-            model       TEXT,
-            input_tokens  INTEGER,
-            output_tokens INTEGER,
-            cost_usd    REAL,
-            created_at  TEXT NOT NULL,
-            FOREIGN KEY (session_id) REFERENCES sessions(id) ON DELETE CASCADE
-         );
-         CREATE INDEX IF NOT EXISTS idx_messages_session ON session_messages(session_id);
-
-         CREATE TABLE IF NOT EXISTS session_tool_calls (
-            id          INTEGER PRIMARY KEY AUTOINCREMENT,
-            session_id  TEXT NOT NULL,
-            message_id  INTEGER,
-            tool_name   TEXT NOT NULL,
-            tool_input  TEXT,
-            tool_output TEXT,
-            status      TEXT NOT NULL DEFAULT 'pending',
-            duration_ms INTEGER,
-            created_at  TEXT NOT NULL,
-            FOREIGN KEY (session_id) REFERENCES sessions(id) ON DELETE CASCADE,
-            FOREIGN KEY (message_id) REFERENCES session_messages(id) ON DELETE SET NULL
-         );
-         CREATE INDEX IF NOT EXISTS idx_tool_calls_session ON session_tool_calls(session_id);
-         CREATE INDEX IF NOT EXISTS idx_tool_calls_name ON session_tool_calls(tool_name);",
-    )
-    .storage_context("failed to initialize session_db schema")?;
-
-    init_fts(conn)?;
-    Ok(())
-}
-
-fn init_fts(conn: &Connection) -> Result<()> {
-    let has_fts: bool = conn
-        .prepare("SELECT 1 FROM sqlite_master WHERE type='table' AND name='sessions_fts'")?
-        .exists([])?;
-
-    if !has_fts {
-        conn.execute_batch(
-            "CREATE VIRTUAL TABLE sessions_fts USING fts5(
-                session_id,
-                agent_definition_name,
-                content,
-                tool_name
-             );",
-        )
-        .storage_context("failed to create sessions_fts virtual table")?;
-    }
-    Ok(())
 }
